chore: Refactor Client and Server classes
Refactor the Client and Server classes to add a new member variable _flagNick and initialize it to false in the constructor. This flag will be used to track if the nickname is already in use. The handleNick function in Server.cpp has been updated to set the flagNick to true when a nickname is already in use. Additionally, the condition in handleNick has been modified to check if the nickname is empty and the flagNick is false before setting the nickname.